refactor: Appropriately handle errors
The general strategy here is to log rusqlite errors with a more helpful
panic message.

In the case of the daemon, we can often recover from a bad MPD command,
in which case we just log and move on.

This is not extensive error handling, but it cleans up 95% of the hack
`unwrap` calls I had scattered around